+++ src/database/dal/receita.dal.ts
@@ -1,7 +1,69 @@
-import { Op, ValidationError } from "sequelize";
+import { Op, ValidationError, WhereOptions } from "sequelize";
 import { GetAllFilters } from "./types";
 import Receita, { ReceitaInput, ReceitaOutput } from "../models/Receita.model";
 import ApiError from "../../error/ApiError";
+import Categoria from "../models/categoria.model";
+
+async function atualizaVlrEmCategoria(
+	categoriaAntiga: string,
+	vlrAntigo: number,
+	categoriaNova: string,
+	vlrNovo: number
+) {
+	let categoria = await Categoria.findByPk(categoriaAntiga)
+		.then((value) => value)
+		.catch((e) => {
+			if (!(e instanceof ValidationError)) {
+				throw e;
+			}
+
+			throw new ApiError({
+				code: 400,
+				err: e,
+			});
+		});
+
+	await (categoria as Categoria)
+		.update({ vlr_em_categoria: categoria.vlr_em_categoria - vlrAntigo })
+		.then((value) => value)
+		.catch((e) => {
+			if (!(e instanceof ValidationError)) {
+				throw e;
+			}
+
+			throw new ApiError({
+				code: 400,
+				err: e,
+			});
+		});
+
+	categoria = await Categoria.findByPk(categoriaNova)
+		.then((value) => value)
+		.catch((e) => {
+			if (!(e instanceof ValidationError)) {
+				throw e;
+			}
+
+			throw new ApiError({
+				code: 400,
+				err: e,
+			});
+		});
+
+	await (categoria as Categoria)
+		.update({ vlr_em_categoria: categoria.vlr_em_categoria + vlrNovo })
+		.then((value) => value)
+		.catch((e) => {
+			if (!(e instanceof ValidationError)) {
+				throw e;
+			}
+
+			throw new ApiError({
+				code: 400,
+				err: e,
+			});
+		});
+}
 
 export const create = async (payload: ReceitaInput): Promise<ReceitaOutput> => {
 	const receita = await Receita.create(payload)
@@ -16,6 +78,14 @@ export const create = async (payload: ReceitaInput): Promise<ReceitaOutput> => {
 				err: e,
 			});
 		});
+
+	atualizaVlrEmCategoria(
+		receita.categoria_id,
+		0,
+		receita.categoria_id,
+		receita.valor
+	);
+
 	return receita;
 };
 
@@ -40,10 +110,13 @@ export const update = async (
 		// @todo throw custom error
 		throw new ApiError({
 			code: 400,
-			err: "Receuta não encontrada!"
+			err: "Receita não encontrada!",
 		});
 	}
-	
+
+	const categoriaAntiga = receita.categoria_id;
+	const valorAntigo = receita.valor;
+
 	const updatedReceita = await (receita as Receita)
 		.update(payload)
 		.then((value) => value)
@@ -57,6 +130,14 @@ export const update = async (
 				err: e,
 			});
 		});
+
+	atualizaVlrEmCategoria(
+		categoriaAntiga,
+		valorAntigo,
+		updatedReceita.categoria_id,
+		updatedReceita.valor
+	);
+
 	return updatedReceita;
 };
 
@@ -73,17 +154,43 @@ export const getById = async (id: string): Promise<ReceitaOutput> => {
 				err: e,
 			});
 		});
+
 	if (!receita) {
 		// @todo throw custom error
 		throw new ApiError({
 			code: 400,
-			err: "Receita não encontrada!"
+			err: "Receita não encontrada!",
 		});
 	}
+
 	return receita;
 };
 
 export const deleteById = async (id: string): Promise<boolean> => {
+	const receita = await Receita.findByPk(id)
+		.then((value) => value)
+		.catch((e) => {
+			if (!(e instanceof ValidationError)) {
+				throw e;
+			}
+
+			throw new ApiError({
+				code: 400,
+				err: e,
+			});
+		});
+
+	if (!receita) {
+		// @todo throw custom error
+		throw new ApiError({
+			code: 400,
+			err: "Receita não encontrada!",
+		});
+	}
+
+	const categoria = receita.categoria_id;
+	const valor = receita.valor;
+
 	const deletedReceitaCount = await Receita.destroy({
 		where: { id },
 	})
@@ -98,15 +205,20 @@ export const deleteById = async (id: string): Promise<boolean> => {
 				err: e,
 			});
 		});
+
+	atualizaVlrEmCategoria(categoria, valor, categoria, 0);
+
 	return !!deletedReceitaCount;
 };
 
 export const getAll = async (
-	filters?: GetAllFilters
+	filters?: GetAllFilters,
+	where?:WhereOptions<Receita>
 ): Promise<ReceitaOutput[]> => {
 	return Receita.findAll({
 		where: {
 			...(filters?.isDeleted && { deletedAt: { [Op.not]: null } }),
+			...(where && where)
 		},
 		...((filters?.isDeleted || filters?.includeDeleted) && {
 			paranoid: true,
